Clean up conversations/ dead code and unnecessary exports
